+++ app/Http/Controllers/SubCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Category;
 use App\SubCategory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use View;
 
 class SubCategoryController extends Controller
@@ -24,6 +25,14 @@ class SubCategoryController extends Controller
 
     public function store(Request $request)
     {
+        $existing = SubCategory::where('name',$request->subcategory_name)->count();
+
+        if($existing > 0)
+        {
+            session()->flash("alert-danger", "Product Subcategory Exists!");
+            return back();
+        }
+
         // dd($request);
         $subcategories = new SubCategory;
         $subcategories->category_id = $request->category_id;
@@ -37,10 +46,18 @@ class SubCategoryController extends Controller
 
     public function update(Request $request)
     {
-     $subcategories = SubCategory::find($request->id);
-     $subcategories->category_id= $request->category_id;
-     $subcategories->name = $request->subcategory_name;
-     $subcategories->save();
+
+        $existing = SubCategory::where('name',$request->subcategory_name)->count();
+
+        if($existing > 0)
+        {
+            session()->flash("alert-danger", "Product Subcategory Exists!");
+            return back();
+        }
+        $subcategories = SubCategory::find($request->id);
+        $subcategories->category_id= $request->category_id;
+        $subcategories->name = $request->subcategory_name;
+        $subcategories->save();
 
       session()->flash("alert-success", "Product Subcategory Updated Successfully!");
         return back();
@@ -48,6 +65,13 @@ class SubCategoryController extends Controller
 
     public function destroy(Request $request)
     {
+        $check_existance = DB::table('inv_products')->where('sub_category_id',$request->id)->count();
+
+        if($check_existance > 0)
+        {
+            session()->flash("alert-danger", "Product Subcategory is in use!");
+            return back();
+        }
         try {
             Subcategory::destroy($request->id);
             session()->flash("alert-danger", "Product Subcategory Deleted successfully!");
